Fix a few trivial float -> int conversions
The implicit conversion leads to a warning starting from python 3.8 and
may be removed in the future.

+++ vimiv/gui/completionwidget.py
@@ -109,7 +109,7 @@ class CompletionView(widgets.FlatTreeView):
         """Resize columnds according to model."""
         for i in range(self.model().columnCount()):
             fraction = self.model().sourceModel().column_widths[i]
-            self.setColumnWidth(i, fraction * self.width())
+            self.setColumnWidth(i, int(fraction * self.width()))
 
 
 def instance():
